fix for chart not loading error
readded async attribute to ow_tracker.html.
Moved calls to functions from top of page to bottom of page so that the charts are not created after being called. This initially worked and probably spend up the page. But this problem likely is due to being able to pull from session storage and not have to make an fetch request every time the page is loaded